% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRNToolSel-package.R
\docType{package}
\name{GRNToolSel-package}
\alias{GRNToolSel}
\alias{GRNToolSel-package}
\title{GRNToolSel: Gene Regulatory Network Tools Selector}
\description{
An R package of Gene Regulatory Network Tools Selector (GRNToolSel) with benchmarking real biological data from high-throughput Omics experiments to compare and assess the performance of various algorithms on gene regulatory network inference. This package contains of several inference algorithms that cover models of Mutual Information, Correlation, Machine learning, Regression algorithms, including microarray and single-cell RNA-seq data of mouse embryonic stem cells. This package provides modules that can be input into gold standards as well as algorithms, which is helpful for the rigorous and scalable evaluation of the GRN construction methods under the specific data.
}
\keyword{internal}
